Added aux.rs, extracted funcs from lib.rs
