+++ native/wasmtime_ex/src/session.rs
@@ -1,5 +1,11 @@
+use lazy_static::lazy_static;
 use std::collections::HashMap;
+use std::sync::RwLock;
 use wasmtime::{Module, Val, ValType};
+
+lazy_static! {
+    pub static ref SESSIONS: RwLock<HashMap<i64, Box<Session>>> = RwLock::new(HashMap::new());
+}
 
 pub struct Session {
     pub module: Module,
